requirements.txt
opencv-python
numpy
imageio
pillow
os